Updated file structure to follow the suggestion my manager told me. It is nicer and keeps components and their respective styles more organized:
