+++ src/components/Header/Header.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import '../stylesheets/header-footer.css'
-import SearchBar from "./SearchBar";
-import SortDropdown from "./SortDropdown";
+import SearchBar from "../SearchBar/SearchBar";
+import SortDropdown from "../SortDropdown/SortDropdown";
 
 export default function Header({handleSearch, handleClear, handleSort}){
