+++ src/components/MovieList/MovieList.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import MovieCard from "./MovieCard";
+import MovieCard from "./MovieCard/MovieCard";
 
 export default function MovieList({movieList}){
 
@@ -7,14 +7,7 @@ export default function MovieList({movieList}){
 
     <>  
       {/*Map through the data array and for each movie data create a movie card and pass the data as prop*/}
-      
-      {movieList.map((movieData) => {
-        return(
-          <MovieCard key = {movieData.id} movie = {movieData}/>
-        )
-      })}
-
-      
+      {movieList.map((movieData) => <MovieCard key = {movieData.id } movie = {movieData}/> )}
     </>
 
   )
